moved stock data functions to separate file and fixed some minor issues
